Added diagnostic for Pixy camera background task
In the event that the Pixy camera background task reads unexpected data
and throws an uncaught exception, a new terminated abnormally flag is
set. Furthermore, a new diagnostic test is added whenever a Pixy
camera is configured which monitors this flag and reports a fault if
it ever becomes true.