Added diagnostic tests for Talon error codes
Talon motor controllers report error codes for commands sent to them
when there is some kind of problem, such as CAN communication issues,
command timeout duration exceeded, etc. In the event that these errors
are frequent (e.g. exceed 100 since a robot was started), this new
diagnostic reports that there might be a problem with the CAN bus.