TrcPixyCam & PixyVision now properly handle frame available with nothing seen
In order for LED status code to determine when the camera has seen nothing,
the TrcPixycam code now returns non-null value when it gets camera regardless
of how many objects it sees (including zero). Furthermore, PixyVision
now caches and returns the last rectangle it calculated from a frame
even if nothing was seen (a rectangle of null).

In the past, there was never any particular interest in this case (we only
really cared about object rectangles when one is in sight), but for showing
a status indicator, this differentiation becomes important.