Improved shutdown hook behavior, improved stopLoggingThread behavior, renamed variables
Now, the JVM hook will wait for the logging thread to finish before giving up control.

Now, when stopping the logging thread, if returnImmediately is true, the program will interrupt and return. If it's false, the program will not interrupt. Instead, it will use the shutdown hook method, where it will set the shutdown flag, which causes the thread to finish writing queued messages and then exit. Then, the shutdown flag is reset to false.